Add definitions for SpreadElementPattern and SpreadPropertyPattern.
Technically these new types are subtypes of Pattern, and so they can be
used wherever a Pattern is expected. It's up to the other Pattern subtypes
to be more specific about what they expect. For instance, ObjectPattern
requires the elements of its .properties field to be either
PropertyPattern or SpreadPropertyPattern nodes, not just any Property
subtype, which makes ObjectPattern a bit safer than ArrayPattern, which
allows its .elements to be any kind of Pattern.