Reference node union type vs string
When `string` is introduced to a union type, we are no longer able
to narrow types because the type widens to `string` rather than a string literal
union.

Here is an example that demonstrates this: http://bit.ly/2QyiplJ

Usage of the unioned node's type is an approach also used by `@babel/types`.
https://gist.github.com/brieb/e312c8f2565c4b2cde5996e1f648780e#file-index-d-ts-L41
My guess is that this was added to address this situation.